+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Track the region as part of the combatant now the combat log includes it.
 
 ### Added
+- [PR 692](https://github.com/aza547/wow-recorder/pull/692/files) - Adds Configurable Push-To-Talk Release Delay.
+
 ### Fixed
 - Make wipes overrun a few seconds longerto avoid missing anything.
 - [Issue 668](https://github.com/aza547/wow-recorder/issues/668) - Fixes the zoom in hotkey.